Integer durations

+++ cfp/views.py
@@ -158,7 +158,7 @@ def get_schedule_generate_csv(request):
                 proposal.proposer.email_addr,
                 proposal.session_type,
                 proposal.title,
-                proposal.length.total_seconds() / 60,
+                int(proposal.length.total_seconds() / 60),
                 '',  # tag eg pydata
                 '',
                 '',
